From review feedback: * Break up theme module * Improved module docs * Rename "dark" to "dark_theme"
